requirements.txt
pychalk>=2.0.1
requests>=2.18.4
urllib3>=1.22
packaging>=19.0